анимашка + порядок в синхорнизации php

+++ delete.php
@@ -1,39 +1,35 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<title>Del</title>
-</head>
-<body>
+<?php 	header('Location:http://examples.cs06028-wordpress.tw1.ru/mindmap/index.php'); ?>
 
 
 
-	<?php 
-	error_reporting(E_ALL & ~E_DEPRECATED); 
-	ini_set('display_errors', 'On'); 
-	restore_error_handler (); 
+
+<?php 
+error_reporting(E_ALL & ~E_DEPRECATED); 
+ini_set('display_errors', 'On'); 
+restore_error_handler(); 
 
 
-	?>
 
 
-	<?php 
 
-	require 'db.php';
+	if(isset($_GET['id']))    #если кнопка нажата
+	{
+		require'db.php';
 
-	$id = $_GET['id'];
+		$id = $_GET['id'];
 
 
-	$del = "DELETE FROM jsmap_db WHERE `id`='$id';";
-	$dbh->query($del);
+		$del = "DELETE FROM jsmap_db WHERE `id`='$id';";
+		$dbh->query($del);
 
 
 	$dbh = NULL;  //Закрыл соединение
-	header('Location: http://host2/index.php');
-	?> 
+	
+}
+?> 
+<script type="text/javascript">
+	// location.replace("http://examples.cs06028-wordpress.tw1.ru/mindmap/index.php");
+</script>
 
 
 
-
-</body>
-</html>
